Move FileDialogCollectionNavigator to its own file (no longer private class) Add class diagram for collectionNavigation
